Чистим, приводим в порядок все после добавления Docking-возможностей.
