+++ DataImpression/Models/ProcessingCore/TobiiCSVRecord.cs
@@ -6,39 +6,33 @@ using System.Threading.Tasks;
 
 namespace DataImpression.Models
 {
+    /// <summary>
+    /// Класс хранит в себе данные одной записи (строки) из csv-файла в формате tobii
+    /// </summary>
     public class TobiiCSVRecord
     {
-       
-            public long time_ms;
-            public List<Column> AOIHitsColumnsInCSVFile = new List<Column>();
-
-            public TobiiCSVRecord()
-            {
-                time_ms = 0;
-            }
-
-            public TobiiCSVRecord(TobiiCSVRecord TR)
-            {
-                time_ms = TR.time_ms;
-            }
-        
-    }
-
-    public class FAOIsOnTimeRecord
-    {
-
+        /// <summary>
+        /// Время в милисекундах
+        /// </summary>
         public long time_ms;
-        public List<FAOI> FAOIs = new List<FAOI>();
 
-        public FAOIsOnTimeRecord()
+        /// <summary>
+        /// Список зон AOI куда испытуемый смотрит в этот момент времени
+        /// </summary>
+        public List<Column> AOIHitsColumnsInCSVFile = new List<Column>();
+
+        public TobiiCSVRecord()
         {
             time_ms = 0;
         }
 
-        public FAOIsOnTimeRecord(FAOIsOnTimeRecord Record)
+        public TobiiCSVRecord(TobiiCSVRecord TR)
         {
-            time_ms = Record.time_ms;
+            time_ms = TR.time_ms;
         }
 
     }
+
+
+    
 }
